refactor: moved the enrichReferences function to a separate file to make it easier to mock
