+++ packages/react-tei/src/SidePanel/bibliographicReferences/useDocumentBibliographicReferences.spec.tsx
@@ -1,11 +1,21 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { describe, expect, it } from "vitest";
+import { describe, expect, it, vi } from "vitest";
 import { renderHook } from "vitest-browser-react";
 import { DocumentContextProvider } from "../../DocumentContextProvider";
 import type { DocumentJson } from "../../parser/document";
+import type { EnrichedReference } from "./enrichReferences";
 import { useDocumentBibliographicReferences } from "./useDocumentBibliographicReferences";
 
-const queryClient = new QueryClient();
+vi.mock("./enrichReferences", () => ({
+	default: (references: DocumentJson[]): Promise<EnrichedReference[]> => {
+		return Promise.resolve(
+			references.map((reference) => ({
+				...reference,
+				validationState: "not_found",
+			})),
+		);
+	},
+}));
 
 function TestWrapper({
 	jsonDocument,
@@ -15,7 +25,7 @@ function TestWrapper({
 	children: React.ReactNode;
 }) {
 	return (
-		<QueryClientProvider client={queryClient}>
+		<QueryClientProvider client={new QueryClient()}>
 			<DocumentContextProvider jsonDocument={jsonDocument}>
 				{children}
 			</DocumentContextProvider>
@@ -65,6 +75,7 @@ describe("useDocumentBibliographicReferences", () => {
 				],
 			},
 		];
+
 		const { result } = await renderHook(
 			() => useDocumentBibliographicReferences(),
 			{
@@ -85,6 +96,7 @@ describe("useDocumentBibliographicReferences", () => {
 			},
 		]);
 	});
+
 	it("should return an empty array if no div[type=references] found", async () => {
 		const jsonDocument = [
 			{
@@ -164,6 +176,7 @@ describe("useDocumentBibliographicReferences", () => {
 				],
 			},
 		];
+
 		const result = await renderHook(
 			() => useDocumentBibliographicReferences(),
 			{
